+++ tools/ZAPD/ZAPD/ZAnimation.cpp
@@ -76,8 +76,8 @@ std::string ZNormalAnimation::GetSourceOutputCode(const std::string& prefix)
 
 		string indicesStr = "";
 		string valuesStr = "    ";
-		const int lineLength = 14;
-		const int offset = 0;
+		const uint8_t lineLength = 14;
+		const uint8_t offset = 0;
 
 		for (size_t i = 0; i < rotationValues.size(); i++)
 		{
@@ -98,12 +98,12 @@ std::string ZNormalAnimation::GetSourceOutputCode(const std::string& prefix)
 		}
 
 		parent->AddDeclarationArray(rotationValuesSeg, DeclarationAlignment::Align16,
-		                            (int)rotationValues.size() * 2, "static s16",
+		                            rotationValues.size() * 2, "static s16",
 		                            StringHelper::Sprintf("%sFrameData", defaultPrefix.c_str()),
 		                            rotationValues.size(), valuesStr);
 
 		parent->AddDeclarationArray(rotationIndicesSeg, DeclarationAlignment::Align16,
-		                            (int)rotationIndices.size() * 6, "static JointIndex",
+		                            rotationIndices.size() * 6, "static JointIndex",
 		                            StringHelper::Sprintf("%sJointIndices", defaultPrefix.c_str()),
 		                            rotationIndices.size(), indicesStr);
 	}
@@ -111,7 +111,7 @@ std::string ZNormalAnimation::GetSourceOutputCode(const std::string& prefix)
 	return "";
 }
 
-int ZNormalAnimation::GetRawDataSize()
+size_t ZNormalAnimation::GetRawDataSize()
 {
 	return 16;
 }
@@ -122,7 +122,7 @@ std::string ZNormalAnimation::GetSourceTypeName()
 }
 
 void ZNormalAnimation::ExtractFromXML(tinyxml2::XMLElement* reader,
-                                      const std::vector<uint8_t>& nRawData, const int nRawDataIndex,
+                                      const std::vector<uint8_t>& nRawData, const uint32_t nRawDataIndex,
                                       const std::string& nRelPath)
 {
 	rawData = std::move(nRawData);
@@ -185,7 +185,7 @@ std::string ZLinkAnimation::GetSourceOutputCode(const std::string& prefix)
 	return "";
 }
 
-int ZLinkAnimation::GetRawDataSize()
+size_t ZLinkAnimation::GetRawDataSize()
 {
 	return 8;
 }
@@ -196,7 +196,7 @@ std::string ZLinkAnimation::GetSourceTypeName()
 }
 
 void ZLinkAnimation::ExtractFromXML(tinyxml2::XMLElement* reader,
-                                    const std::vector<uint8_t>& nRawData, const int nRawDataIndex,
+                                    const std::vector<uint8_t>& nRawData, const uint32_t nRawDataIndex,
                                     const std::string& nRelPath)
 {
 	rawData = std::move(nRawData);
@@ -210,8 +210,6 @@ void ZLinkAnimation::ParseRawData()
 	ZAnimation::ParseRawData();
 
 	const uint8_t* data = rawData.data();
-
-	// segmentAddress = GETSEGOFFSET(BitConverter::ToInt32BE(data, rawDataIndex + 4));
 	segmentAddress = (BitConverter::ToInt32BE(data, rawDataIndex + 4));
 }
 
@@ -286,7 +284,7 @@ void ZCurveAnimation::ParseRawData()
 }
 
 void ZCurveAnimation::ExtractFromXML(tinyxml2::XMLElement* reader,
-                                     const std::vector<uint8_t>& nRawData, const int nRawDataIndex,
+                                     const std::vector<uint8_t>& nRawData, const uint32_t nRawDataIndex,
                                      const std::string& nRelPath)
 {
 	ZResource::ExtractFromXML(reader, nRawData, nRawDataIndex, nRelPath);
@@ -427,7 +425,7 @@ void ZCurveAnimation::PreGenValues(const std::string& prefix)
 	}
 }
 
-int ZCurveAnimation::GetRawDataSize()
+size_t ZCurveAnimation::GetRawDataSize()
 {
 	return 0x10;
 }
