+++ tools/ZAPD/ZAPD/ZBackground.h
@@ -11,16 +11,16 @@ protected:
 
 public:
 	ZBackground(ZFile* nParent);
-	ZBackground(const std::string& prefix, const std::vector<uint8_t>& nRawData, int nRawDataIndex,
+	ZBackground(const std::string& prefix, const std::vector<uint8_t>& nRawData, uint32_t nRawDataIndex,
 	            ZFile* nParent);
 	void ParseRawData() override;
 	void ParseBinaryFile(const std::string& inFolder, bool appendOutName);
 	void ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-	                    int nRawDataIndex, const std::string& nRelPath) override;
+	                    uint32_t nRawDataIndex, const std::string& nRelPath) override;
 
 	void CheckValidJpeg(const std::string& filepath);
 
-	int GetRawDataSize() override;
+	size_t GetRawDataSize() override;
 
 	void DeclareVar(const std::string& prefix, const std::string& bodyStr);
