+++ tools/ZAPD/ZAPD/ZMtx.h
@@ -12,13 +12,13 @@ protected:
 public:
 	// ZMtx() = default;
 	ZMtx(ZFile* nParent);
-	ZMtx(const std::string& prefix, const std::vector<uint8_t>& nRawData, int nRawDataIndex,
+	ZMtx(const std::string& prefix, const std::vector<uint8_t>& nRawData, uint32_t nRawDataIndex,
 	     ZFile* nParent);
 	void ParseRawData() override;
 	void ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-	                    int nRawDataIndex, const std::string& nRelPath) override;
+	                    uint32_t nRawDataIndex, const std::string& nRelPath) override;
 
-	int GetRawDataSize() override;
+	size_t GetRawDataSize() override;
 
 	void DeclareVar(const std::string& prefix, const std::string& bodyStr);
