+++ tools/ZAPD/ZAPD/ZRoom/Commands/SetObjectList.cpp
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-SetObjectList::SetObjectList(ZRoom* nZRoom, std::vector<uint8_t> rawData, int rawDataIndex)
+SetObjectList::SetObjectList(ZRoom* nZRoom, std::vector<uint8_t> rawData, uint32_t rawDataIndex)
 	: ZRoomCommand(nZRoom, rawData, rawDataIndex)
 {
 	objects = vector<uint16_t>();
@@ -33,7 +33,7 @@ string SetObjectList::GenerateExterns()
 	                             segmentOffset);
 }
 
-string SetObjectList::GenerateSourceCodePass1(string roomName, int baseAddress)
+string SetObjectList::GenerateSourceCodePass1(string roomName, uint32_t baseAddress)
 {
 	string sourceOutput = "";
 
@@ -61,7 +61,7 @@ string SetObjectList::GenerateSourceCodePass1(string roomName, int baseAddress)
 	return sourceOutput;
 }
 
-int32_t SetObjectList::GetRawDataSize()
+size_t SetObjectList::GetRawDataSize()
 {
 	return ZRoomCommand::GetRawDataSize() + (objects.size() * 2);
 }
