+++ tools/ZAPD/ZAPD/ZSymbol.h
@@ -7,22 +7,19 @@ class ZSymbol : public ZResource
 {
 protected:
 	std::string type;
-	uint32_t typeSize;
+	size_t typeSize;
 	bool isArray = false;
 	uint32_t count = 0;
 
 public:
 	ZSymbol(ZFile* nParent);
-	// ZSymbol(const std::string& nName, int nRawDataIndex, const std::string& nType, uint32_t
-	// nTypeSize, bool nIsArray, uint32_t nCount); ZSymbol(tinyxml2::XMLElement* reader, const
-	// std::vector<uint8_t>& nRawData, int nRawDataIndex, ZFile* nParent);
 
 	void ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-	                    const int nRawDataIndex, const std::string& nRelPath) override;
+	                    const uint32_t nRawDataIndex, const std::string& nRelPath) override;
 
 	void ParseXML(tinyxml2::XMLElement* reader) override;
 
-	int GetRawDataSize() override;
+	size_t GetRawDataSize() override;
 
 	std::string GetSourceOutputHeader(const std::string& prefix) override;
