+++ tools/ZAPD/lib/assimp/include/assimp/Bitmap.h
@@ -65,7 +65,7 @@ protected:
 
     struct Header {
         uint16_t type;
-        uint32_t size;
+        size_t size;
         uint16_t reserved1;
         uint16_t reserved2;
         uint32_t offset;
@@ -81,7 +81,7 @@ protected:
     };
 
     struct DIB {
-        uint32_t size;
+        size_t size;
         int32_t width;
         int32_t height;
         uint16_t planes;
